Re-enable zlib
zlip-data.js contains a lot of code which was passed to eval(), making it
less than ideal for Hermes. Printed it, formatted it with prettier
created a dummy InitializeZlibBenchmark().

Finally, added a dummy `function read()` which is expected to be
present, but is not used.

This allowed me to re-enable zlip in run-noeval.js.